Move Command struct to new module
