+++ src/hyperfine/benchmark.rs
@@ -5,6 +5,7 @@ use std::process::{ExitStatus, Stdio};
 use colored::*;
 use statistical::{mean, median, standard_deviation};
 
+use crate::hyperfine::command::Command;
 use crate::hyperfine::format::{format_duration, format_duration_unit};
 use crate::hyperfine::internal::{get_progress_bar, max, min, MIN_EXECUTION_TIME};
 use crate::hyperfine::outlier_detection::{modified_zscores, OUTLIER_THRESHOLD};
@@ -12,7 +13,7 @@ use crate::hyperfine::shell::execute_and_time;
 use crate::hyperfine::timer::wallclocktimer::WallClockTimer;
 use crate::hyperfine::timer::{TimerStart, TimerStop};
 use crate::hyperfine::types::{
-    BenchmarkResult, CmdFailureAction, Command, HyperfineOptions, OutputStyleOption,
+    BenchmarkResult, CmdFailureAction, HyperfineOptions, OutputStyleOption,
 };
 use crate::hyperfine::units::Second;
 use crate::hyperfine::warnings::Warnings;
